requirements.txt
alabaster==0.7.10
alembic==0.9.7
ansible==2.4.2.0
apipkg==1.4
asn1crypto==0.22.0
attrs==17.3.0
Automat==0.6.0
autopep8==1.3.2
Babel==2.5.0
backports==1.0
bcrypt==3.1.4
bleach==2.1.1
blinker==1.4
cached-property==1.3.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
constantly==15.1.0
cov-core==1.15.0
coverage==4.2
cracklib==2.9.6
cryptography==2.1.4
Cython==0.26.1
decorator==4.1.2
defusedxml==0.5.0
docker==2.7.0
docker-compose==1.18.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dominate==2.3.1
entrypoints==0.2.3
execnet==1.5.0
extras==1.0.0
file-magic==0.3.0
fixtures==3.0.0
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gdbus-codegen==2.52.3
gemato==10.3
gentoolkit==0.4.2
get==1.0.2
gevent==1.2.2
gpg==1.10.0
greenlet==0.4.12
guess-language-spirit==0.5.3
hgdistver==0.25
html5lib==0.999999999
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.6.1
ipyparallel==6.0.2
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
java-config==2.2.0
jedi==0.10.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.4.0
layman==2.4.2
linecache2==1.0.0
LiveWires==2.1
lxml==4.1.1
m2r==0.1.12
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==0.23
meson==0.44.0
micawber==0.3.4
mimeparse==1.6.0
mistune==0.7.4
mysql-connector-python==2.1.7
nbconvert==5.2.1
nbformat==4.4.0
ndg-httpsclient==0.4.2
netaddr==0.7.19
netlink==1.0
notebook==5.0.0
numpy==1.13.3
packaging==16.8
pandocfilters==1.4.2
paramiko==2.3.1
path.py==8.1.2
pathlib2==2.3.0
pbr==3.1.1
peewee==2.10.2
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
ply==3.10
portage==2.3.25
post==1.0.1
prompt-toolkit==1.0.15
psutil==5.4.1
ptyprocess==0.5.2
public==1.0.1
pwquality==1.4.0
py==1.4.34
pyasn1==0.3.7
pyasn1-modules==0.1.5
pyblake2==1.1.0
pycairo==1.15.4
pycodestyle==2.3.1
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
pyelftools==0.24
pygame==1.9.3
Pygments==2.2.0
pygobject==3.24.1
PyJWT==1.6.1
PyNaCl==1.2.1
pyOpenSSL==17.4.0
pyparsing==2.2.0
pyrsistent==0.13.0
PySocks==1.6.7
pytest==3.2.2
pytest-cache==1.0
pytest-cov==2.5.1
pytest-pep8==1.0.6
pytest-timeout==1.2.0
python-dateutil==2.6.1
python-dotenv==0.8.2
python-editor==1.0.3
python-mimeparse==1.6.0
python-subunit==1.2.0
pytz==2017.2
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
query-string==1.0.1
regex==2017.4.5
request==1.0.1
requests==2.18.4
rfc3987==1.3.7
scandir==1.6
service-identity==17.0.0
setuptools-scm==1.15.6
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.15
ssl-fetch==0.4
strict-rfc3339==0.7
terminado==0.6
testpath==0.1
testtools==2.3.0
texttable==0.9.1
tornado==4.5.1
traceback2==1.4.0
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.2
unittest2==1.1.0
urllib3==1.22
vcversioner==2.16.0.0
virtualenv==15.1.0
visitor==0.1.3
vsearch==1.0
wcwidth==0.1.7
webcolors==1.5
webencodings==0.5.1
websocket-client==0.37.0
Werkzeug==0.14.1
Whoosh==2.7.4
wstools==0.4.5
WTForms==2.1
zope.interface==4.4.3
